chore(person360): the deliverability section takes its own file under the length cap
